Saint lague model
Created class for holding data about a party and also created the main program that uses this data to calculate seats allocated to parties.
Created valuable methods for extracting the data about the parties to be used in calculation. Still need to finish main to allow users to run the program.